% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{explore}
\alias{explore}
\title{Explore a rijkspalette}
\usage{
explore(x, ...)
}
\arguments{
\item{x}{A rijkspalette object}

\item{...}{other arguments passed to plot function}
}
\description{
This function plots the available colours in ab-space along with the cluster
centroids used to generate the palette. By manipulating the number of
clusters, each distinct colour category can receive its own centroid. This
function was designed to work in \code{RStudio}.
}
