% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{imgToLabmat}
\alias{imgToLabmat}
\title{Image to small lab coordinate matrix}
\usage{
imgToLabmat(img)
}
\arguments{
\item{img}{cimg object from imager}
}
\description{
Translates an image into a palette of 9 colours
}
\keyword{internal}
