% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{imgToPalette}
\alias{imgToPalette}
\title{Image to palette}
\usage{
imgToPalette(img, k, brightness)
}
\arguments{
\item{img}{cimg object from imager}

\item{k}{how many colours}

\item{lightness}{how light the returned palette should be}
}
\description{
Translates an image into a palette of 9 colours
}
\keyword{internal}
