% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{labmatToPalette}
\alias{labmatToPalette}
\title{Labmat to palette}
\usage{
labmatToPalette(labmat, k, lightness)
}
\arguments{
\item{labmat}{a n*3 matrix signifying Lab space coordinates}

\item{k}{how many colours}

\item{lightness}{how light the returned palette should be}
}
\description{
Translates a labmat to a colour palette
}
\keyword{internal}
