% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.labmat}
\alias{plot.labmat}
\title{Plot labmat}
\usage{
\method{plot}{labmat}(x, k = 5, ...)
}
\arguments{
\item{x}{a labmat object}

\item{k}{number of clusters to draw}

\item{...}{other arguments passed to plot function}
}
\description{
Plots a labmat object from a rijkspalette. It can be used for tuning the
required number of clusters.
}
\seealso{
\code{\link{explore}}
}
