% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.rgbcols}
\alias{plot.rgbcols}
\title{Plot rijkspalette colours}
\usage{
\method{plot}{rgbcols}(x, ...)
}
\arguments{
\item{x}{an rgbcols object}

\item{...}{other arguments passed to plot function}
}
\description{
Plots a cols object from a rijkspalette.
}
\seealso{
\code{\link{plot.rijkspalette}}
}
