% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.rijkspalette}
\alias{plot.rijkspalette}
\title{Plot rijkspalette}
\usage{
\method{plot}{rijkspalette}(x, ...)
}
\arguments{
\item{x}{rijkspalette object}

\item{...}{other arguments passed to plot function}
}
\description{
Plots a rijkspalette with its source image
}
