% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{setBrightness}
\alias{setBrightness}
\title{Set brightness for a rijkspalette}
\usage{
setBrightness(rp, brightness = 0.5)
}
\arguments{
\item{rp}{rijkspalette object}

\item{brightness}{between 0 (darkest) and 1 (brightest)}
}
\description{
Proper method stuff will be there soon.
}
