% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tune}
\alias{tune}
\title{Tune a rijksPalette}
\usage{
tune(x, lightness, k, ...)
}
\arguments{
\item{x}{rijkspalette object}

\item{lightness}{overall lightness between 0 (darkest) and 1 (brightest)}

\item{k}{number of colours to extract}

\item{...}{other arguments passed to tune function (not used)}
}
\description{
This function tunes the extracted palette from an image in a rijksPalette
object. Often, the defaults need to be tuned a little to get the nicest
results.
}
