% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{explore}
\alias{explore}
\alias{explore.rijkspalette}
\title{Explore the available colours in a rijkspalette}
\usage{
\method{explore}{rijkspalette}(x, ...)

\method{explore}{rijkspalette}(x, ...)
}
\arguments{
\item{x}{A rijkspalette object}

\item{...}{other arguments passed to plot function}
}
\description{
This function plots the available colours in ab-space along with the chosen
cluster centroids. By manipulating the number of clusters, each distinct
colour category can receive its own centroid. This function was tested to
works in RStudio. The package `manipulate` is required for its function.
}
