% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.rijkspalette}
\alias{print.rijkspalette}
\title{Print rijkspalette}
\usage{
\method{print}{rijkspalette}(x, ...)
}
\arguments{
\item{x}{rijkspalette object}

\item{...}{other arguments to print (not used)}
}
\description{
Prints a rijkspalette object with nice colours in the console
}
\seealso{
\code{\link{plot.rijkspalette}}

\code{\link{tune}}

\code{\link{explore}}
}
