% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rijkspalette.R
\name{rijksPalette}
\alias{rijksPalette}
\title{Rijkspalette function}
\usage{
rijksPalette(query)
}
\arguments{
\item{query}{Keyword to search for in the collection of the Rijksmuseum}
}
\value{
an object of type \code{rijkspalette}
}
\description{
This function queries the Rijksmuseum collection for paintings and returns
a colour palette based on colours from that painting.
}
\examples{
\donttest{
pal <- rijksPalette("Vermeer Letter")
plot(pal)
barplot(1/sqrt(1:15), col = pal$palette(15))
}

}
\seealso{
\code{\link{tune}}

\code{\link{explore}}
}
